requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data Validation & Settings
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Earth Engine
earthengine-api==0.1.382
google-auth==2.23.4
google-cloud-storage==2.10.0

# Geospatial Libraries
geopandas==0.14.1
shapely==2.0.2
rasterio==1.3.9
pyproj==3.6.1

# Image Processing
pillow==10.1.0
numpy>=1.24.4

# HTTP Client
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# System Monitoring
psutil==5.9.6
